server/tests: Create separate downloadable fixture
