+++ server/tests/benefit/benefits/test_downloadables.py
@@ -1,116 +1,24 @@
-from collections.abc import Sequence
-from typing import Any, cast
 from uuid import UUID
 
 import pytest
-from sqlalchemy.orm import contains_eager
 
-from polar.benefit.benefits.downloadables import BenefitDownloadablesService
 from polar.benefit.schemas import BenefitDownloadablesCreateProperties
 from polar.file.schemas import FileRead
 from polar.models import (
-    Benefit,
     Downloadable,
-    File,
     Organization,
     Product,
     User,
 )
-from polar.models.benefit import BenefitDownloadables, BenefitType
 from polar.models.downloadable import DownloadableStatus
-from polar.models.subscription import SubscriptionStatus
-from polar.postgres import AsyncSession, sql
+from polar.postgres import AsyncSession
 from tests.fixtures.database import SaveFixture
-from tests.fixtures.random_objects import (
-    create_benefit,
-    create_benefit_grant,
-    create_subscription,
-)
+from tests.fixtures.downloadable import TestDownloadable
 
 
 @pytest.mark.asyncio
 @pytest.mark.http_auto_expunge
 class TestDownloadblesBenefit:
-    async def create_benefit_and_grant(
-        self,
-        session: AsyncSession,
-        save_fixture: SaveFixture,
-        user: User,
-        organization: Organization,
-        product: Product,
-        properties: BenefitDownloadablesCreateProperties,
-    ) -> tuple[BenefitDownloadables, dict[str, Any]]:
-        benefit = await create_benefit(
-            save_fixture,
-            type=BenefitType.downloadables,
-            organization=organization,
-            properties=properties.model_dump(mode="json"),
-        )
-        return await self.create_grant(
-            session,
-            save_fixture,
-            cast(BenefitDownloadables, benefit),
-            user=user,
-            product=product,
-        )
-
-    async def create_grant(
-        self,
-        session: AsyncSession,
-        save_fixture: SaveFixture,
-        benefit: BenefitDownloadables,
-        user: User,
-        product: Product,
-    ) -> tuple[BenefitDownloadables, dict[str, Any]]:
-        subscription = await create_subscription(
-            save_fixture,
-            product=product,
-            user=user,
-            status=SubscriptionStatus.active,
-        )
-        await create_benefit_grant(
-            save_fixture,
-            user,
-            benefit,
-            subscription=subscription,
-        )
-        return await self.run_grant_task(session, benefit, user)
-
-    async def run_grant_task(
-        self, session: AsyncSession, benefit: BenefitDownloadables, user: User
-    ) -> tuple[BenefitDownloadables, dict[str, Any]]:
-        service = BenefitDownloadablesService(session)
-        granted = await service.grant(benefit, user, {})
-        return benefit, granted
-
-    async def run_revoke_task(
-        self, session: AsyncSession, benefit: BenefitDownloadables, user: User
-    ) -> tuple[BenefitDownloadables, dict[str, Any]]:
-        service = BenefitDownloadablesService(session)
-        revoked = await service.revoke(benefit, user, {})
-        return benefit, revoked
-
-    async def get_user_downloadables(
-        self, session: AsyncSession, user: User
-    ) -> Sequence[Downloadable]:
-        statement = (
-            sql.select(Downloadable)
-            .join(File)
-            .join(Benefit)
-            .options(contains_eager(Downloadable.file))
-            .where(
-                Downloadable.user_id == user.id,
-                File.deleted_at.is_(None),
-                File.is_uploaded == True,  # noqa
-                File.is_enabled == True,  # noqa
-                Benefit.deleted_at.is_(None),
-            )
-        )
-
-        res = await session.execute(statement)
-        downloadables = res.scalars().all()
-        return downloadables
-
     @pytest.mark.auth
     async def test_grant_one(
         self,
@@ -121,7 +29,7 @@ class TestDownloadblesBenefit:
         product: Product,
         uploaded_logo_jpg: FileRead,
     ) -> None:
-        _, granted = await self.create_benefit_and_grant(
+        _, granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -133,7 +41,7 @@ class TestDownloadblesBenefit:
         )
         assert granted["files"][0] == str(uploaded_logo_jpg.id)
 
-        downloadables = await self.get_user_downloadables(session, user)
+        downloadables = await TestDownloadable.get_user_downloadables(session, user)
         assert downloadables
         assert len(downloadables) == 1
 
@@ -158,7 +66,7 @@ class TestDownloadblesBenefit:
             uploaded_logo_png,
         ]
 
-        _, granted = await self.create_benefit_and_grant(
+        _, granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -169,7 +77,7 @@ class TestDownloadblesBenefit:
             ),
         )
 
-        downloadables = await self.get_user_downloadables(session, user)
+        downloadables = await TestDownloadable.get_user_downloadables(session, user)
         assert downloadables
         assert len(downloadables) == len(files)
 
@@ -191,7 +99,7 @@ class TestDownloadblesBenefit:
         uploaded_logo_jpg: FileRead,
         uploaded_logo_png: FileRead,
     ) -> None:
-        _, granted = await self.create_benefit_and_grant(
+        _, granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -210,7 +118,7 @@ class TestDownloadblesBenefit:
         assert len(granted["files"]) == 1
         assert granted["files"][0] == str(uploaded_logo_png.id)
 
-        downloadables = await self.get_user_downloadables(session, user)
+        downloadables = await TestDownloadable.get_user_downloadables(session, user)
         assert downloadables
         assert len(downloadables) == 1
 
@@ -228,7 +136,7 @@ class TestDownloadblesBenefit:
         product: Product,
         uploaded_logo_jpg: FileRead,
     ) -> None:
-        benefit, granted = await self.create_benefit_and_grant(
+        benefit, granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -245,7 +153,7 @@ class TestDownloadblesBenefit:
         assert len(granted["files"]) == 1
         assert granted["files"][0] == str(uploaded_logo_jpg.id)
 
-        downloadables = await self.get_user_downloadables(session, user)
+        downloadables = await TestDownloadable.get_user_downloadables(session, user)
         assert downloadables
         assert len(downloadables) == 1
 
@@ -253,10 +161,12 @@ class TestDownloadblesBenefit:
         assert downloadable.status == DownloadableStatus.granted
         assert downloadable.file_id == uploaded_logo_jpg.id
 
-        await self.run_revoke_task(session, benefit, user)
+        await TestDownloadable.run_revoke_task(session, benefit, user)
 
         # Now revoked
-        updated_downloadables = await self.get_user_downloadables(session, user)
+        updated_downloadables = await TestDownloadable.get_user_downloadables(
+            session, user
+        )
         assert updated_downloadables
         assert len(updated_downloadables) == 1
 
@@ -279,7 +189,7 @@ class TestDownloadblesBenefit:
             uploaded_logo_jpg,
             uploaded_logo_png,
         ]
-        benefit, granted = await self.create_benefit_and_grant(
+        benefit, granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -292,17 +202,21 @@ class TestDownloadblesBenefit:
 
         # First granted
         assert len(granted["files"]) == 2
-        granted_downloadables = await self.get_user_downloadables(session, user)
+        granted_downloadables = await TestDownloadable.get_user_downloadables(
+            session, user
+        )
         assert len(granted_downloadables) == 2
         for i, file in enumerate(files):
             grant = granted_downloadables[i]
             assert grant.file_id == file.id
             assert grant.status == DownloadableStatus.granted
 
-        await self.run_revoke_task(session, benefit, user)
+        await TestDownloadable.run_revoke_task(session, benefit, user)
 
         # Now revoked
-        revoked_downloadables = await self.get_user_downloadables(session, user)
+        revoked_downloadables = await TestDownloadable.get_user_downloadables(
+            session, user
+        )
         assert len(revoked_downloadables) == 2
         for i, file in enumerate(files):
             revoked = revoked_downloadables[i]
@@ -330,7 +244,7 @@ class TestDownloadblesBenefit:
             },
             files=[f.id for f in files],
         )
-        benefit, granted = await self.create_benefit_and_grant(
+        benefit, granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -342,7 +256,7 @@ class TestDownloadblesBenefit:
         assert len(granted["files"]) == 1
         assert granted["files"][0] == str(files[1].id)
 
-        downloadables = await self.get_user_downloadables(session, user)
+        downloadables = await TestDownloadable.get_user_downloadables(session, user)
         assert downloadables
         assert len(downloadables) == 1
 
@@ -355,10 +269,12 @@ class TestDownloadblesBenefit:
         session.add(benefit)
         await session.flush()
 
-        _, updated_granted = await self.run_grant_task(session, benefit, user)
+        _, updated_granted = await TestDownloadable.run_grant_task(
+            session, benefit, user
+        )
 
         assert len(updated_granted["files"]) == 2
-        downloadables = await self.get_user_downloadables(session, user)
+        downloadables = await TestDownloadable.get_user_downloadables(session, user)
         assert downloadables
         assert len(downloadables) == 2
 
@@ -392,7 +308,7 @@ class TestDownloadblesBenefit:
             uploaded_logo_jpg,
             uploaded_logo_png,
         ]
-        benefit, user_granted = await self.create_benefit_and_grant(
+        benefit, user_granted = await TestDownloadable.create_benefit_and_grant(
             session,
             save_fixture,
             user=user,
@@ -405,7 +321,9 @@ class TestDownloadblesBenefit:
 
         # First user granted all files
         assert len(user_granted["files"]) == 2
-        user_downloadables = await self.get_user_downloadables(session, user)
+        user_downloadables = await TestDownloadable.get_user_downloadables(
+            session, user
+        )
         assert len(user_downloadables) == 2
         for i, file in enumerate(files):
             grant = user_downloadables[i]
@@ -422,7 +340,7 @@ class TestDownloadblesBenefit:
 
         # Second user granted one file
         # Since they subscribe after the 2nd file was archived
-        _, user_second_granted = await self.create_grant(
+        _, user_second_granted = await TestDownloadable.create_grant(
             session,
             save_fixture,
             benefit,
@@ -430,7 +348,7 @@ class TestDownloadblesBenefit:
             product=product,
         )
         assert len(user_second_granted["files"]) == 1
-        user_second_downloadables = await self.get_user_downloadables(
+        user_second_downloadables = await TestDownloadable.get_user_downloadables(
             session, user_second
         )
         assert len(user_second_downloadables) == 1
